NFC: Add LSAN leak suppression to ASAN pipeline (#7503)
Address-sanitizer reports a leak coming from the dynamic shared library
loading code, though dlopen() is paired with dlclose(). This leak
doesn't manifest on main yet because dxc adds dxcompiler directly to
target_link_libraries (which it shouldn't have to), and dxil is loaded
within that library. A PR to remove dxil loading from dxcompiler exposes
the leak when dxc loads dxil dynamically.

This change adds a suppressions file for LSAN to suppress the leak in
the loader code that happens under call_init. This also changes the
Linux_Clang_Release build to RelWithDebInfo so symbols are present.