Require complete types in some missing places
This change adds two additional calls to Sema::RequireCompleteType, one
when evaluating unary `sizeof`, which fixes a reported issue with
templates. The second is in the DXR diagnostics where complete type
should be required but not diagnosed because ordering in the compiler is
a bit wonky when diagnosing DXR entry points.

Fixes #7510

../tools/clang/test/SemaHLSL/raytracing-entry-diags.hlsl
../tools/clang/test/SemaHLSL/sizeof-requires-complete-type.hlsl