Require complete types in some missing places (#7511)
This change adds two additional calls to Sema::RequireCompleteType, one
when evaluating unary `sizeof`, which fixes a reported issue with
templates. The second is in the DXR diagnostics where complete type
should be required but not diagnosed because ordering in the compiler is
a bit wonky when diagnosing DXR entry points.

Fixes #7510

---------

Co-authored-by: Tex Riddell <texr@microsoft.com>