[SPIRV] Use unknown image format in vk1.3 and later
We have had many request to use the `unknown` image format for storage
images (OpTypeImage with sampled=2). We did not want to do that when
targeting earlier versions of Vulkan because it could break existing
code. The capability StorageImageWriteWithoutFormat is not guarenteed to
be available on Vulkan 1.1 devices. This means the application will stop
working.

However, Vulkan 1.3 guarentees that StorageImageWriteWithoutFormat and
StorageImageReadWithoutFormat are available. We can make this change for
VK1.3 and later without breaking existing code.

Fixes #7478