Long Vector Execution tests preliminary work to merge a minor refactor. (#7532)
This resolves issue #7531. This is a preliminary step to merging in some
of the Long Vector Execution tests currently sitting in the
staging-sm6.9 branch. There are no functional changes here, but given
that this is a refactor I do not want to add the [NFC] tag to the PR
title.

- Moves functions used by existing exec tests and incoming long vector
tests to a common HlslExecTestUtils.h.
- Updates naming to adhere to LLVM coding standards for newly created
files (even though the functions aren't new)
- Move a few other shared functions to files that make more sense than
ExecutionTest.cpp.
- TableParameterHandler class moved to its own header/cpp files. No
naming updates as nothing else was touched. Can update in a subsequent
PR if it is preferred.
- Add the LLVM coding guidelines preferred include guard to
HlslTestUtils.h to mitigate redefinition issues exposed by this
refactor.
- Updated the D3D shader model 'redefines' in ExecutionTest.cpp as they
were also factored out into a common header. constexpr required because
they are enum values.
- BigObj added to the cmake file as I was hitting issues locally for
this.