[SPIRV] Emit DebugScope in wrapper (#7341) (#7529)
Legalization and optimization will produce inaccurate NS100 debug info
if there is no DebugScope emitted in the wrapper function. This PR
corrects this oversight and renames the wrapper to "__dxc_setup". This
may cause a stack frame named __dxc_setup to appear in an HLSL debugger,
however, users should be familiar with this type of thing. A C debugger
might show crt0, or a debugger can filter this frame out of the user's
view. This PR addresses
[7341](https://github.com/microsoft/DirectXShaderCompiler/issues//7341)