Revert "[SPIRV] Use unknown image format in vk1.3 and later" (#7572)
I just learned more about the VK feature. In VK1.3, the validation rule
was moved from the existence of the capability to being specific to the
format. It is possible that people will see regressions if their code
runs on a driver that does not support
VK_FORMAT_FEATURE_2_STORAGE_WRITE_WITHOUT_FORMAT_BIT or
VK_FORMAT_FEATURE_2_STORAGE_READ_WITHOUT_FORMAT_BIT for the format used
by the developer.

Reverts microsoft/DirectXShaderCompiler#7528