[SPIR-V] Explicitly state which layout rules require scalar block layout (#7539)
I was trying to debug a Vulkan Storage Buffer-related memory alignment
issue in my application where I was using SPIR-V generated via `dxc`
with `-fvk-use-dx-layout`. In `SPIR-V.rst`, I happened to miss the
paragraph that follows the list of layout rules (removed in this
proposal). That paragraph starts with "To use scalar layout", which
given my use of DirectX layout, I did not think was relevant to me.
However, the next sentence of that paragraph sneakily and indirectly
mentions that `VK_EXT_scalar_block_layout` is required for the DirectX
memory layout as well.

I have proposed explicitly stating the extension requirement when the
relevant layout rules are listed.