[SPIRV] Allow spirv type as template parameter (#7626)
SPIR-V intrinsics allow us to create spirv basic type and opaque type in
HLSL, but these type are object and not allowed in template parameter.

```fundamental
error: object 'Int8Type' is not allowed in builtin template parameters
    /* OpTypeCooperativeMatrixKHR */ 4456, Int8Type,
                                           ^
```

This doesn't make sense to me, and is not convenience to use. This
change wants to allow that use those in template parameter.