[SPIRV] Add option to use the Unknown image format (#7632)
Many Vulkan driver allow the SPIR-V to use the `Unknown` image format
for storage images and texel buffers. This makes the SPIR-V more
flexible, and it useful for developers. However, it can be cumbersome to
have to add the `vk::image_format` attribute to all resources.

This option allows users to make `Unkown` the default image format for
all resource types.

Fixes #7484