[SPIRV] Fix DebugSource for files which are not found (#7662)
https://github.com/microsoft/DirectXShaderCompiler/issues/4218 and
https://github.com/microsoft/DirectXShaderCompiler/pull/4362 identified
and partially fixed problems with DebugSource generation.
https://github.com/microsoft/DirectXShaderCompiler/pull/6085 recognized
that exe and API users were having different outputs in DebugSource, and
attempted to fix that. However, as the conversation in
https://github.com/microsoft/DirectXShaderCompiler/issues/7569 shows,
the solution was not entirely correct. This PR causes dxc to connect the
input source string, whether by exe or API, to one and only one
DebugSource, like other compilers. Files which are not found produce a
DebugSource without the second operand.
This PR fixes
https://github.com/microsoft/DirectXShaderCompiler/issues/7569