Remove `hlsl` namespace from SemaHLSL (#7656)
HLSL intrinsics were previously being added to the translation unit decl
but assigned a namespace decl context which was then used only in
codegen to determine if special handling was needed for the namespace.
This causes an odd array of behaviors including strange mangling and
confusing diagnostics.

This PR removes the HLSL namespace and properly adds intrinsics to
namespaces when appropriate for the `vk` and `dx` namespaces. This
should allow qualified and unqualified name lookup to behave as
expected, correct odd mangling, and improve the confusing diagnostics.

The PR introduces a new HLSLBuiltinCallAttr to replace the use of the
hlsl namespace. This attribute denotes when a function declaration
should be emitted in codegen as an HLSL builtin. It is subtly different
from HLSLIntrinsicAttr which is attached to all intrinsic functions and
allows the AST to track the HL OpCode.

---------

Co-authored-by: Ashley Coleman <ascoleman@microsoft.com>