Implement basic AsType tests for long vectors. (#7668)
This PR adds test cases for most of the 'AsType' HLSL intrinsics. It
introduces changes to facilitate testing an HLSL intrinsic with an
output type that does not match the input type. This functionality
required some changes to the TestConfig object as well as the addition
of a 'VariantVector' alias to store the expected output.

Verified locally by running new and all existing long vector exec test
cases against WARP.

Resolves #7473