Allow native vector intrinsics with fixed component types (#7597)
Several intrinsics that support native vectors return the same element
type regardless of overload. For native vectors, these need to return a
vector of size matching the overload with the fixed element type. This
provides the source generation to allow it and the changes to the
related lowering operations to generate the correct DXIL.

This includes some queries that return boolean information about the
vector components in a boolean vector output and also the unary bit
operations which operate on integers and return integers containing
information about the component bits.

Incidentally moves some intrinsics that have long since not been
scalarized out of the scalarized test. They were updated to expect the
new output, but were in the wrong location and not tested as thoroughly
as their native vector counterparts.

Fixes #7342

---------

Co-authored-by: Greg Roth <grroth@microsoft.com>
Co-authored-by: Ashley Coleman <ascoleman@microsoft.com>