Refactor long vector exec tests to use classes for the op types and some (#7679)
This PR refactors the long vector execution tests I have been working on
to remove the need for a LongVectorOpTypeT template parameter and
instead rely on OOP/Polymorphism. This is a much cleaner approach for
this application. I Initially chose not to go that route thinking that
the templated path for the LongVectorOpTypeT would be clean and easy.
That proved to be wrong as I added additional ops. Also gets rid of the
tpp file as it's no longer required.

**Edit:** Added an additional commit which re-factors some of the test
setup a little more by introducing a struct for test metadata that helps
simplify the code in each test op type constructor.