[SM6.9] Scalarize native vector intrinsics on linking pre6.9 (#7690)
When shaders are compiled into libraries as 6.9+ they need to be
downgraded for previous shader models upon linking. This adapts the
load/store scalarization pass that serves this purpose for those
operations to work for any DXIL intrinsics that need it.

Incidentally clarifies the role of and tightens up the testing of vector
intrinsics that are still scalarized for now.

Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7344