Execution Tests: Fix bug parsing bools in TableParameterHandler (#7701)
This PR fixes a bug in the Exec tests utility TableParamaterHandler
which parses TAEF metadata which has either been populated via an xml
file or TAEF macros. There was a bug when parsing a TableParameter::BOOL
a string was being passed to TryGetValue instead of a bool.

The first argument to TryGetValue is the name of the parameter and the
second is a reference to the variable to store the value.

How tested: I'm adding a test that's using the handler to retrieve a
bool value. I confirmed that it's working now.