Create new class to handle external validation, and rename existing dll loading class. (#7514)
This PR creates a class that is responsible for loading and unloading
dxil.dll, allowing for external validation.
Tests are added to test this class's functionality. Additionally, an
interface is defined that joins the existing singular dll loading helper
class, and the new class that tracks two dlls. The classes were renamed
to be more abstract, as well as all existing uses of the original
singular dll loader class.
Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7422

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>