[SPIR-V] Fix invalid codegen for empty cbuffers
Each variable in the Uniform/Uniform Constant storage class must be
decorated with DescriptorSet + Binding.
-> VUID-StandaloneSpirv-UniformConstant-06677

When a cbuffer was empty (or only contained resources), an empty
struct and variable in the uniform storage class was created.
Because the cbuffer was empty, no binding/descriptor set was assigned,
hence no decoration was emitted.

This was fine as long as we compiled with optimizations on, as DCE would
get rid of this unused cbuffer. But compiling with -O0 would yield a
validation error.

Fixes #7681