Add back compat lit test suite, and download older dxil.dll's (#7705)
This PR adds new lit test suites, that build by downloading and
unzipping 5 release packages. The release packages are a selection of
historical packages that will be useful in later testing for validation
back-compatibility. However, this data can also be used for other
testing, and doesn't necessarily need to be restricted to older
packages. For now, 5 releases have been included, including the latest 3
releases at the time of this writing.
This PR then executes these test suites with an environment variable
that points to the dxil.dll that was just downloaded, so that in the
future, DxcDllExtValLoader can grab the dxil.dll path and initialize
appropriately.
This PR effectively sets up a test suite and an environment so that back
compat work is possible.
Adds some final touches on @damyanp's work.
Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7707

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>