[Rebase] Execution Tests: Long Vector tests add Binary operator tests (#7757)
This PR adds tests cases for the remaining binary operators. I also
opted to include a single unary bitwise operator because it made sense
to logically group the bitwise operators.

Validated new tests cases against latest WARP. There is currently an
issue leading to failures with <, >, <=, and >= on int64 values that
does not appear to be a test bug.

Resolves #7464

---------

Co-authored-by: Damyan Pepper <damyanp@microsoft.com>