Use WARP from nuget by default (#7427)
Previously using WARP from nuget was something that had to be opted in
to. This change makes it the default behvior.

This allows some simplifications. The nuget package is installed at
configure time. If the user wants to override the nuget installation
behavior then they can pass in extra parameters, rather than having an
alternative mechanism for describing which package to use.

The actual DLL to load is specified through a TAEF parameter and
specifies the full path to the DLL.