Refactor LongVectors.cpp to improve binary sizes (#7776)
This moves some things around so that less code is generated at compile
time and some information about data types and operations are looked up
at runtime.

For a debug build this reduces the size of the DLL by about 1.9MB, and
the PDB by about 8.1MB.

It also allows us to remove the `/bigobj` compiler flag.