Add requires dxil for tests with shader models above 1.6 (#7796)
There are quite a few tests that are missing a `REQUIRES: dxil-1-<minor
version>` at the beginning of the file. Because of this, when the test
is run with a 1.6 external validator, the tests fail. The tests should
be treated as unsupported if the target is above 1.6.
This PR adds this filter to many of the tests that fail with a 1.6
validator.
Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7793