Execution Tests: Add test cases for Select (#7798)
This PR adds test cases for 'select'. 
Select is a ternary op where the first argument is a bool vector. The
2nd and 3rd argument types must match.
I opted to keep all inputs as the same type to make implementation of
the test easier in our framework. Instead, the shader code converts the
first input to a bool value via a conditional expression.

New test cases validated against a private version of WARP.

Completion of this PR resolves #7468