[SPIRV] Add diagnostic for boolean bitfields
Adds a diagnostic to flag the use of boolean bitfields in structs when targeting
SPIR-V. This is not supported for variables that are not externally visible.

This change introduces a new error message in SemaHLSL to detect and report this
unsupported usage for:
- Local variables
- Function parameters
- Static global variables

A new test file is added to verify the diagnostic is triggered in these cases
and not for externally visible global variables.

Avoid the segfault reported in
https://github.com/microsoft/DirectXShaderCompiler/issues/7790.