[SPIR-V] Fix incorrect [branch] gen on return. (#7834)
PR #6756 tried to address an issue when variables were declared in a
fallthrough, but contained a bug: the PR assumed return statements were
leaves, and that no statement was associated with the node. This is not
true if a value is returned as the return statements contains the
expression describing the returned value.

Skipping this return statement means we'd generate invalid code when
returning a value in a switch with the `[branch]` attribute. Moved the
exit condition after recording the statement so early returns are still
generated.

Fixes #7832