Make DXC and DXV use new `DxcDllExtValidationLoader` to load external validator before validation. (#7749)
This PR focuses on allowing dxc to use the existing infrastructure to
use a `DxcDllExtValidationLoader` object, and load an external dxil.dll
with it, via the environment variable. The validate*.test was added to
verify that the dll is indeed being loaded and used for external
validation, which causes a failure, since the dll's validator version is
older than the required minimum validator version that the metadata
indicates.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>