Execution Tests: Long Vector - Dot fp32 precision (#7801)
Update Dot expected values to be computed using doubles to improve
precision of the final result.
Additionally, this PR adds logic to compute an absolute epsilon
tolerance for the final result of each Dot operation by summing absolute
epsilons calculated on each multiply and addition in the dot product.
The summation portion is re-ordered to a worst-case ordering when
computing the absolute epsilon error tolerance as the DX spec allows for
this.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Tex Riddell <texr@microsoft.com>