[SPIR-V] Fix resource heap & fvk-bind-register interactions
The interaction between `-fvk-bind-register` and the resource/sampler
heaps were not tested, and broken.

This commit solves this by requiring the `-fvk-bind-sampler-heap` or
`-fvk-bind-resource-heap` flag to be used if the `-fvk-bind-register`
flag is present and a resource/sampler heap is present.

An alternative solution would be to allow implicit bindings on heaps
while other resources require the register attribute, but I think it
would be less useful as the goal of using this flag is to have full
control over the bindings.

This commit refactorizes a bit the helper functions
isResourceHeap/isSamplerHeap, but this part is NFC. The main bit
is in DeclResultIdMapper, where a special case is added the the heaps.

Fixes #7857