[NFC] Clean up HlslExecTestUtils
An attempt was made to factor out some common code from ExecutionTest.cpp to be shared between LongVectors.cpp and ExecutionTest.cpp. Somehow this either wasn't completed, or got lost when it was merged from the 6.9 staging branch.

This change finishes extracting this code, as well as making it clear what the HlslExecTestUtils interface is.