Remove hctgettaef.py, fix link to WDK download page
hctgettaef.py downloaded a very old, unsupported version of TAEF.  We should be guiding people to download official supported versions, such as the one that you get from the Windows Driver Kit.

hctdeploytest.cmd was using hctgettaef.py, but as I can't find any references to it anywhere I'm going to assume it is unused and will remove it rather than try to maintain it.