Added debug info for localized global variables for inlined scopes. (#7799)
Previously, global variables converted to `alloca`'s, their debug info
is written as a `dbg.value` for a fake local variable prefixed with
"global.". This was only done for the entry function's `DISubprogram`,
which makes the variable inaccessible for all inlined function scopes.

This change creates a `dbg.value` for each inlined `DISubprogram`, each
pointing to a new `DILocalVariable` with the inlined `DISubprogram` as
its scope.