Fix double-increment bug in LegalizeResources when looping over PHIs
This subtle bug can cause a failure to resolve a static resource in a
corner case where DxilValueCache could replace an odd-indexed phi, and this
is needed to resolve other phis to known values, but an indexing bug
prevents the loop from ever visiting any phi nodes at odd-numbered indexes
in the list.

This change also adds two IR-based pass tests for this code path, as there
weren't any before.