LongVectors tests: Fix device leak and improve logging (#7951)
Previously, if a device is removed then the code that recreates it
wouldn't release the old device pointer, resulting in the device
leaking. This actually prevents removed device recovery entirely, since
all references to a removed device need to be released before a new one
can be created. This change fixes that.

Also:

* improve logging in situations where requirements aren't met (eg trying
to create a SM 6.9 device on a system that doesn't support SM 6.9)
* return false from class setup if we couldn't create a device - this
will cause all the tests in the class to be skipped or failed (depending
on the FailIfRequirementsNotMet setting) - reducing log noise, and
saving the time that would be taken to run tests that we know are going
to fail
* improve logging when recreating a device so we can tell the difference
between "device was removed" and "there wasn't a device set"
* add a warning log if createDevice is called with something that looks
like it already has a value set


Fixes #7949