Make ExecutionTests able to pick a specific AgilitySDK (#7963)
The original AgilitySDK API only allowed paths relative to the process's
executable to be used to specificy the location of the AgilitySDK. So,
previously, if we wanted to use a specific AgilitySDK with the execution
tests we had to arrange for the AgilitySDK's dlls to be placed in a
subdirectory underneath TAEF's te.exe. This required a convoluted dance
involving finding and copying TAEF executables somewhere that could be
owned by the testing process.

This change makes it so we can use
[ID3D12DeviceFactory](https://microsoft.github.io/DirectX-Specs/d3d/IndependentDevices.html)
- which allows the AgilitySDK location to be specified using absolute
paths.

A new class, D3D12SDKSelector, is added to help manage this, as well as
falling back to using the old-style AgilitySDK and the inbox D3D12.

---------

Co-authored-by: github-actions[bot] <github-actions[bot]@users.noreply.github.com>
Co-authored-by: Helena Kotas <hekotas@microsoft.com>