Add separate experimental DXIL op table using high OpCode bit (#7947)
Implements:
https://github.com/microsoft/hlsl-specs/blob/main/proposals/0052-experimental-dxil-ops.md

This change adds support for an experimental DXIL operation table, which
provides an independent numeric space from the main OpCode set.

It adds support for indexing into OpCode tables using the high 16-bits
as the table index, and the low 16-bits as the index into the OpCode
table.

While the fundamental change could support more tables for individual
experimental features or extensions, it's currently limited to tables
`0` for `CoreOps` and `0x8000` for `ExperimentalOps` without additional
work to unlock more tables. This maps to the existing opcodes for
`CoreOps` and opcodes with the high bit set for `ExperimentalOps`.

Now, db_dxil_op_table manages dxil op lists, and DXIL ops are built
using table methods.