Execution Tests: Update tolerances for half precision trig functions (#7992)
This PR adjusts the tolerances for half precision trig functions with
DXIL vectors. The D3D spec only notes the tolerance range for fp32. That
fp32 error tolerance is not sufficient for halfs. After some further
discussion we opted to use 0.003 as the tolerance for most half
precision trig ops. For trig ops with a range that approaches infinity
in either direction, we instead use 2 ULP.

Fixes #7936