Fix CodeQL security issues (#7985)
This path fixes 7 issues raised by codeql:

- `dxcapi.use.h` - [SM01925] Uncontrolled process operation: Add
comment, so the scan can ignore this issue.
- `ProgramTest.cpp` - [SM01932] User-controlled data may not be null
terminated: Add comment, so the scan can ignore this issue.
- `DxbcConverter.cpp` - [SM01928] Comparison of narrow type with wide
type in loop condition: Change affected variable to be `unsigned`
instead of `BYTE`.
- `ExecutionTest.cpp` - [SM01733] Too few arguments to formatting
function: Remove an argument that didn't seem to be used.
- `DxbcUtil.cpp` - [SM01928] Comparison of narrow type with wide type in
loop condition: Cast the result of `+` operator back to BYTE, since it
was implicit being cast to an int.
- `DSAclean.py` - [SM03905] Inefficient regular expression: Script
doesn't seem to be used anymore, so it is being removed.
- `CaptureCmd` - [SM02167] Weak hashes : Script doesn't seem to be used
anymore, so it is being removed.

---------

Co-authored-by: Joao Saffran <jderezende@microsoft.com>