Execution Tests: Fix W4505 (unreferenced local function has been removed) offenses (#8003)
Some downstream consumers of the exec tests are required to enable
[W4505](https://learn.microsoft.com/en-us/cpp/error-messages/compiler-warnings/compiler-warning-level-4-c4505?view=msvc-170)
as an error.
This is triggered when a static function is removed via dead code
removal for not being used in its translation unit. The recommended fix
for this is to instead declare the shared functions in the header as
inline.