Fix incorrect OpAccessChain for inherited struct members (#8005)
Adds a new test case to verify that the OpAccessChain is generated
correctly for inherited members of a struct.

This fixes an issue where the indices for a base class were being
prepended to the indices of the derived class, resulting in an incorrect
access chain.

Fixes https://github.com/microsoft/DirectXShaderCompiler/issues/7968