Add HL lower table validation and opcode json dependency (#8004)
It's easy to get gLowerTable out of order in HLOperationLower.cpp. That
can result in lowering an HLSL intrinsic to the wrong DXIL op.

This change adds static_assert validation for the opcode table, which
will report the specific index where the opcode didn't match.

Validation caught two out-of-order ops in the table, which are also
fixed in this change.

The change also adds hlsl_intrinsic_opcodes.json to dependencies in
add_hlsl_hctgen, since the `hlsl::IntrinsicOp` enum is dependent on it.