[SPIR-V] Fix void on extented inst sets
Inline SPIR-V on normal instructions had a condition to avoid
emitting a result type on instruction with a void return type.
This is because instructions like `OpStore` have a void return type, but
in SPIR-V have no result id.

OpExtInst when defining void instruction (like debugBreak()) always have
a result id, even with a void result type. This commits fixes the
condition to reflect this.

Fixes #8012